winSAS-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
reportlab
Pillow
pypdf